Remove Clojure dependencies from being counted as outdated
Remove Clojure dependencies before counting stats on whether they are up
to date or not.

There are lots of reasons why a project doesn't have the latest Clojure
version, and there is nothing particularly wrong with that dep being out
of date, as a user's own Clojure will take precedence over the one in
the dep.

- Update Clojure to 1.8
- Add sift function to simplify the stats calculation code

Fixes #28